+++ packages/react-native-editor/CHANGELOG.md
@@ -11,9 +11,32 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 
+## 1.74.0
+
+-   [**] [Quote block] Adds support for V2 behind a feature flag [#40133]
+-   [**] Update "add block" button's style in default editor view. [#39726]
+-   [*] Remove banner error notification on upload failure [#39694]
+
+## 1.73.1
+
+-   [*] [Spacer block] Fix crash when changing the height value using the text input [#40053]
+
+## 1.73.0
+
+-   [*] Update react-native-reanimated version to 2.4.1 [#39430]
+-   [*] Upgrade Gradle to 7.4 & AGP to 7.1.1 [#39508]
+-   [*] Add waits to fix editor test flakiness [#39668]
+
+## 1.72.1
+
+-   [*] Detect GIF badge during render [#39882]
+
+## 1.72.0
+
 -   [*] Add GIF badge for animated GIFs uploaded to Image blocks [#38996]
 -   [*] Small refinement to media upload errors, including centring and tweaking copy. [#38951]
 -   [*] Update gesture handler and reanimated libraries [#39098]
+-   [*] Fix issue with list's starting index and the order [#39354]
 
 ## 1.71.3